Updates: Neo4j 4.4, neo4j-graphql-java 1.5.0, graphql 16.2
The changes to neo4j-graphql-java has caused the transpiler to become much stricter with validation. There still seem to be some problems there which is possibly the cause that we cannot use a transpiler interceptor anymore to change the query there. The workaround implemented now is to perform cypher query transformation after transpilation from graphql to cypher. This is now encapsulated in the new TimeVersioningGraphQLToCypherTranslator class.